+++ USB_API/USB_Common/usbdma.c
@@ -86,8 +86,8 @@ void * memcpyDMA (void * dest, const void *  source, size_t count)
     //DMA4 workaround - disable DMA transfers during read-modify-write CPU 
     //operations
     DMA_disableTransferDuringReadModifyWrite();
-    DMA_setSrcAddress(USB_DMA_CHAN, (uint32_t)source, DMA_DIRECTION_INCREMENT);
-    DMA_setDstAddress(USB_DMA_CHAN, (uint32_t)dest, DMA_DIRECTION_INCREMENT);
+    DMA_setSrcAddress(USB_DMA_CHAN, (uint32_t)((uintptr_t)source), DMA_DIRECTION_INCREMENT);
+    DMA_setDstAddress(USB_DMA_CHAN, (uint32_t)((uintptr_t)dest), DMA_DIRECTION_INCREMENT);
     //DMA4 workaround - re-enable DMA transfers during read-modify-write CPU 
     //operations
     DMA_enableTransferDuringReadModifyWrite();
